[REFACTOR] : step1-index 로직을 controller 폴더 파일로 분리
