+++ src/controller/game.js
@@ -1,0 +1,37 @@
+import LottoResult from "../domain/LottoResult.js";
+import LottoStatus from "../domain/LottoStatus.js";
+import Ticket from "../domain/Ticket.js";
+import Output from "../view/Output.js";
+import { OUTPUT } from "../constants/message.js";
+import { PRICE } from "../constants/price.js";
+import { divideByUnit } from "../utils/count.js";
+import { getNeededLottoNumbers, getPrice } from "./getInputWithRetry.js";
+
+const game = async () => {
+  const price = await getPrice();
+
+  const countNumber = divideByUnit(PRICE.UNIT, price);
+  Output.print(`${countNumber}${OUTPUT.BUY_COUNT}`);
+
+  const lottos = Ticket.createLottos(countNumber);
+  Output.printLottoNumber(lottos);
+
+  const { winningLotto, bonusLottoNumber } = await getNeededLottoNumbers();
+  Output.print(OUTPUT.WINNING_HISTORY);
+  Output.print(OUTPUT.LINE);
+
+  const lottoStatus = new LottoStatus({
+    issuedLottoNumbers: lottos.map((lotto) => lotto.getLottoNumbers()),
+    enteredLottoNumbers: winningLotto.getLottoNumbers(),
+    bonusLottoNumber,
+  });
+  const matchedlottoStatus = lottoStatus.getMatchedLottoStatus();
+
+  const lottoResult = new LottoResult(matchedlottoStatus, price);
+  const lottoHistory = lottoResult.getWinningHistory();
+
+  Output.printWinningHistory(lottoHistory);
+  Output.printTotalProfit(lottoResult.getRate());
+};
+
+export default game;
